app/requirements.txt
Flask-Cors==3.0.9
Flask-WTF==0.14.3
python-dateutil==2.8.1
numpy==1.19.1
eventlet==0.27.0
requests==2.24.0
numexpr==2.7.1
pandas==1.1.1
pymongo==3.11.0
